requirements.txt
django==3.2.10
django-extensions==3.1.4
djangorestframework==3.11.2
jsonschema==4.17.3
psycopg2==2.8.3
werkzeug==0.16.0
